remove local testing for switchboard

+++ programs/dsla/src/instructions/deploy_sla.rs
@@ -94,7 +94,7 @@ pub fn handler(
     messenger_address: Pubkey,
     leverage: DslaDecimal,
     start: u128,
-    n_periods: usize,
+    n_periods: u32,
     period_length: PeriodLength,
 ) -> Result<()> {
     let sla = &mut ctx.accounts.sla;
